fix: update get started and get involved pages
