--- src/components/pages/home/discover/index.js
@@ -1,1 +1,0 @@
-export { default } from './discover';
